Xong module User

+++ common/paging.go
@@ -5,7 +5,7 @@ import "strings"
 type Paging struct {
 	Page       int    `json:"page" form:"page"`
 	Limit      int    `json:"limit" form:"limit"`
-	Total      int    `json:"total" form:"total"`
+	Total      int64  `json:"total" form:"total"`
 	FakeCursor string `json:"cursor" form:"cursor"`
 	NextCursor string `json:"next_cursor" form:"next_cursor"`
 }
